+++ BidMachineAdMobAdapter/BidMachineAdMobAdapter/AdProviders/Model/CustomParamsKey.swift
@@ -1,0 +1,12 @@
+//
+//  CustomParamsKey.swift
+//  BidMachineAdMobAdapter
+//
+//  Created by Dzmitry on 23/12/2024.
+//
+
+import Foundation
+
+enum CustomParamsKey {
+    static let mediationMode = "mediation_mode"
+}